Atualizando codigo fonte para versão do dia 19/08

+++ codigo-fonte/src/main/java/com/eurolearn/repository/UsuarioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.eurolearn.models.UsuarioModel;
 
-public interface UsuarioRepository extends JpaRepository<UsuarioModel, Integer> {
-
+public interface UsuarioRepository extends JpaRepository<UsuarioModel, Long> {
+	
 }
